Seperated VerhaalDetail into components and increased font-size
