+++ examen-verhaal-frontend/src/components/story/BackButton.jsx
@@ -1,0 +1,31 @@
+import { useNavigate } from "react-router-dom";
+
+const BackButton = () => {
+  const navigate = useNavigate();
+
+  return (
+    <div className="mb-8">
+      <button
+        onClick={() => navigate("/verhalen")}
+        className="flex items-center text-gray-600 hover:text-gray-800 transition-colors"
+      >
+        <svg
+          className="w-5 h-5 mr-2"
+          fill="none"
+          stroke="currentColor"
+          viewBox="0 0 24 24"
+        >
+          <path
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            strokeWidth={2}
+            d="M10 19l-7-7m0 0l7-7m-7 7h18"
+          />
+        </svg>
+        Terug naar verhalen
+      </button>
+    </div>
+  );
+};
+
+export default BackButton;
